+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Remove IDDFS search algorithm and related parameters.
+- Remove deprecated look-ahead branching heuristic.
 
 ## [0.2.4] - 2026-07-17